+++ packages/contract/src/common/multicall.ts
@@ -1,0 +1,2 @@
+export { Multicall } from '@openzeppelin/contracts/utils/Multicall.sol'
+export { Multicall3 } from './Multicall3.sol'