+++ packages/contract/src/index.ts
@@ -1,5 +1,6 @@
 export type { TevmContract } from './TevmContract'
 export { createTevmContract } from './createTevmContract'
+export { createTevmContractFromAbi } from './createTevmContractFromAbi'
 // reexported because they are needed to be able to
 // instanciate an TevmContract with a json abi
 export { formatAbi, parseAbi } from 'abitype'